refactor: ディレクトリ構造を変更
cloud funtionにdeployするために構造を変更しました。
