+++ .gitignore
@@ -20,5 +20,4 @@
 # Go workspace file
 go.work
 
-.vscode/
 verification.html
