+++ run-test.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+cd adapter
+go test
